Update index.md completed
I completed my update - taking into account the read-me, the notes in the columns of the excel template as well as the general recommendations (beyond protected areas) of Balthasar's tutorial. 

The links to the two illustrations were lost - need to be linked again. 

I also added a new recommendation on filling the GADM GID, Level and version in the "Project location identifier" column, which needs to be discussed. I also already mentioned Kobo Toolbox, because this should be ready during the next days.  

And I mentioned the geodata collection ToR everywhere where information by the KfW counterpart may be required or desired. We can then keep all info to be provided by KfW in these ToR. I will upload the updated ToR also in Github.