Fixes to the layout of kobo install (#39)
* Let's start to see the changes on the references (#3)

* Create How-to_for_KfW_staff

* Rename How-to_for_KfW_staff to How-to_for_project_location_collectors.md

* Update index.md

Only the table is not finished yet. Need to check consistency of text on complex geometries and DAC/CRS-codes.

* Create How-to-for-KfW-Counterparts

This file explains to KfW Conuterparts

* Delete docs/How-to_for_project_location_collectors.md

* Update index.md

including first half of the table until Column J.

* Update index.md

* Update index.md

* Update index.md

* Update index.md completed

I completed my update - taking into account the read-me, the notes in the columns of the excel template as well as the general recommendations (beyond protected areas) of Balthasar's tutorial. 

The links to the two illustrations were lost - need to be linked again. 

I also added a new recommendation on filling the GADM GID, Level and version in the "Project location identifier" column, which needs to be discussed. I also already mentioned Kobo Toolbox, because this should be ready during the next days.  

And I mentioned the geodata collection ToR everywhere where information by the KfW counterpart may be required or desired. We can then keep all info to be provided by KfW in these ToR. I will upload the updated ToR also in Github.

* Update index.md small corrections

* Update index.md small corrections

* Update index.md small corrections

---------

Co-authored-by: Maja4Dev <113998721+Maja4Dev@users.noreply.github.com>

* Further updates by Maja (#4)

* Create How-to_for_KfW_staff

* Rename How-to_for_KfW_staff to How-to_for_project_location_collectors.md

* Update index.md

Only the table is not finished yet. Need to check consistency of text on complex geometries and DAC/CRS-codes.

* Create How-to-for-KfW-Counterparts

This file explains to KfW Conuterparts

* Delete docs/How-to_for_project_location_collectors.md

* Update index.md

including first half of the table until Column J.

* Update index.md

* Update index.md

* Update index.md

* Update index.md completed

I completed my update - taking into account the read-me, the notes in the columns of the excel template as well as the general recommendations (beyond protected areas) of Balthasar's tutorial. 

The links to the two illustrations were lost - need to be linked again. 

I also added a new recommendation on filling the GADM GID, Level and version in the "Project location identifier" column, which needs to be discussed. I also already mentioned Kobo Toolbox, because this should be ready during the next days.  

And I mentioned the geodata collection ToR everywhere where information by the KfW counterpart may be required or desired. We can then keep all info to be provided by KfW in these ToR. I will upload the updated ToR also in Github.

* Update index.md small corrections

* Update index.md small corrections

* Update index.md small corrections

* Delete docs/How-to-for-KfW-Counterparts

---------

Co-authored-by: Maja4Dev <113998721+Maja4Dev@users.noreply.github.com>

* Create How-To_for_KfW_staff

This serves as primary orientation for KfW staff on how to request Project Location Data Collection with minimum effort.

* Update and rename How-To_for_KfW_staff to howto_for_kfw_staff.md

* Update mkdocs.yml

* Fixes to the layout

* Clean out old stuff

---------

Co-authored-by: Maja4Dev <113998721+Maja4Dev@users.noreply.github.com>
Co-authored-by: Surafel <91615452+S-r-f-l@users.noreply.github.com>